requirements.txt
django == 3.2
djangorestframework
pillow
django-cors-headers==3.10.0